Added possibility to take the attendance.
Created new class Attendance that is capturing the attendance.
That is done in menu by user input to all members that are currently in the database.
So there were also minor changes in other classes to comply with these changes.
There is temporary possibility to see the result of the attendace gathering, but it will be removed once the save attendance sheet is implemented.